refactor(jpa): move base entities to "~.support"
