+++ core-modules/nettee-jpa-core/src/main/java/nettee/jpa/support/BaseTimeEntity.java
@@ -1,4 +1,4 @@
-package nettee.jpa;
+package nettee.jpa.support;
 
 import jakarta.persistence.EntityListeners;
 import jakarta.persistence.MappedSuperclass;
@@ -7,12 +7,13 @@ import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
+import java.io.Serializable;
 import java.time.Instant;
 
 @Getter
 @MappedSuperclass
 @EntityListeners(AuditingEntityListener.class)
-public abstract class UuidBaseTimeEntity extends UuidBaseEntity {
+public abstract class BaseTimeEntity implements Serializable {
     @CreatedDate
     private Instant createdAt;
