+++ core-modules/nettee-jpa-core/src/main/java/nettee/jpa/support/LongBaseTimeEntity.java
@@ -1,4 +1,4 @@
-package nettee.jpa;
+package nettee.jpa.support;
 
 import jakarta.persistence.EntityListeners;
 import jakarta.persistence.MappedSuperclass;
@@ -7,13 +7,12 @@ import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-import java.io.Serializable;
 import java.time.Instant;
 
 @Getter
 @MappedSuperclass
 @EntityListeners(AuditingEntityListener.class)
-public abstract class BaseTimeEntity implements Serializable {
+public abstract class LongBaseTimeEntity extends LongBaseEntity {
     @CreatedDate
     private Instant createdAt;
