+++ core-modules/nettee-jpa-core/src/main/java/nettee/jpa/support/UuidBaseTimeEntity.java
@@ -1,4 +1,4 @@
-package nettee.jpa;
+package nettee.jpa.support;
 
 import jakarta.persistence.EntityListeners;
 import jakarta.persistence.MappedSuperclass;
@@ -12,7 +12,7 @@ import java.time.Instant;
 @Getter
 @MappedSuperclass
 @EntityListeners(AuditingEntityListener.class)
-public abstract class LongBaseTimeEntity extends LongBaseEntity {
+public abstract class UuidBaseTimeEntity extends UuidBaseEntity {
     @CreatedDate
     private Instant createdAt;
